# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: open-kraken HTTP API
  version: 0.1.0
  description: |
    Browser-facing HTTP contract for the open-kraken migration workspace.
    WebSocket handshake and event schemas are documented in x-websocket and in
    docs/api/http-websocket-contract.md.
servers:
  - url: http://localhost:8080
paths:
...
  /api/workspaces/{workspaceId}/terminals:
    get:
      operationId: listTerminals
      summary: List terminal sessions
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Terminal session summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTerminalsResponse'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      operationId: createTerminal
      summary: Create or reuse a terminal session
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTerminalRequest'
      responses:
        '200':
          description: Terminal session created or reused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTerminalResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
...
  /api/v1/skills:
    get:
      operationId: listSkills
      summary: List filesystem skill catalog entries
      responses:
        '200':
          description: Skill catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/v1/members/{memberId}/skills:
    get:
      operationId: listMemberSkills
      summary: List skills assigned to a member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member skill bindings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberSkillsResponse'
        '500':
          $ref: '#/components/responses/Error500'
    put:
      operationId: replaceMemberSkills
      summary: Replace the full skill binding list for a member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceMemberSkillsRequest'
      responses:
        '200':
          description: Updated member skill bindings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberSkillsResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error:
                  code: auth.unauthorized
                  message: bearer token is missing or invalid
                  status: 401
                  requestId: req_01KN9B00000000000000000003
                  retryable: false
                  details:
                    workspaceId: ws_123
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbiddenScope:
              value:
                error:
                  code: auth.capability_denied
                  message: member cannot dispatch to this terminal
                  status: 403
                  requestId: req_01KN9B00000000000000000004
                  retryable: false
                  details:
                    workspaceId: ws_123
                    memberId: member_member
                    requiredCapability: terminal.dispatch
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error422:
      description: Unprocessable entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error503:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message, status, requestId, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: integer
        requestId:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
...
    RecoveryState:
      type: object
      required: [mode, resyncRequired, terminalReplay, dedupeKey]
      properties:
        mode:
          type: string
          enum: [snapshot, replay, snapshot_resync]
        lastAckCursor:
          type: [string, 'null']
        resyncRequired:
          type: boolean
        terminalReplay:
          type: string
          enum: [none, snapshot_only, delta_after_snapshot]
        dedupeKey:
          type: string
          const: cursor_then_terminal_seq
...
    Skill:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        path:
          type: string
        category:
          type: string
        contentSummary:
          type: string
    SkillListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
...
    MemberSkillsResponse:
      type: object
      required: [memberId, skills]
      properties:
        memberId:
          type: string
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
    ReplaceMemberSkillsRequest:
      type: object
      required: [skills]
      properties:
        skills:
          type: array
          items:
            type: string
...
    TerminalSnapshotBuffer:
      type: object
      required: [data, rows, cols, cursorRow, cursorCol]
      properties:
        data:
          type: string
        rows:
          type: integer
        cols:
          type: integer
        cursorRow:
          type: integer
        cursorCol:
          type: integer
        history:
          type: string
    TerminalSnapshot:
      type: object
      required: [terminalId, seq, buffer]
      properties:
        terminalId:
          type: string
        seq:
          type: integer
        buffer:
          $ref: '#/components/schemas/TerminalSnapshotBuffer'
    TerminalSession:
      type: object
      required:
        [terminalId, workspaceId, terminalType, command, status, seq, unackedBytes, keepAlive, recovery, createdAt, updatedAt, snapshot]
      properties:
        terminalId:
          type: string
        memberId:
          type: string
        workspaceId:
          type: string
        terminalType:
          type: string
        command:
          type: string
        status:
          type: string
          enum: [created, starting, connecting, online, working, offline, closed, failed]
        seq:
          type: integer
        unackedBytes:
          type: integer
        keepAlive:
          type: boolean
        recovery:
          $ref: '#/components/schemas/RecoveryState'
        metadata:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        snapshot:
          $ref: '#/components/schemas/TerminalSnapshot'
    ListTerminalsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TerminalSession'
    CreateTerminalRequest:
      type: object
      required: [terminalType, command, cols, rows, keepAlive]
      properties:
        memberId:
          type: string
        terminalType:
          type: string
        command:
          type: string
        cols:
          type: integer
        rows:
          type: integer
        keepAlive:
          type: boolean
    CreateTerminalResponse:
      type: object
      required: [session]
      properties:
        session:
          $ref: '#/components/schemas/TerminalSession'
...